Attempting to get Graphics

+++ include/LinearRegression.h
@@ -13,13 +13,13 @@ namespace Regression
 	class Linear
 	{
 	public:
-		Linear(std::vector<int> X, std::vector<int> Y);
-		float predict(int X, char entryType);
+		Linear(std::vector<float> X, std::vector<float> Y);
+		float predict(float X, char entryType);
 		float getA();
 		float getB();
-		float calcualteValue(int val);
+		float calcualteValue(float val);
 	private:
-		void train(std::vector<int> X, std::vector<int> Y);
+		void train(std::vector<float> X, std::vector<float> Y);
 		float a, b;
 	};
 }
