+++ src/LinearRegression.cpp
@@ -1,11 +1,11 @@
 #include "LinearRegression.h"
 
-Regression::Linear::Linear(std::vector<int> X, std::vector<int> Y)
+Regression::Linear::Linear(std::vector<float> X, std::vector<float> Y)
 {
     train(X, Y);
 }
 
-float Regression::Linear::predict(int X, char entryType)
+float Regression::Linear::predict(float X, char entryType)
 {
     if (entryType == Xentry)
     {
@@ -27,12 +27,12 @@ float Regression::Linear::getB()
     return b;
 }
 
-float Regression::Linear::calcualteValue(int val)
+float Regression::Linear::calcualteValue(float val)
 {
     return 0.0f;
 }
 
-void Regression::Linear::train(std::vector<int> X, std::vector<int> Y)
+void Regression::Linear::train(std::vector<float> X, std::vector<float> Y)
 {
     float sumX = 0, sumX2 = 0, sumY = 0, sumYX = 0;
